fix: 修复标签删除404错误（类型转换问题）
## 问题根因
- 多选删除标签时返回404错误
- 后端`db.First(&tag, tagID)`中`tagID`是`uint64`类型
- 但GORM的First方法期望`uint`类型，导致类型不匹配
- 查询失败，返回404"标签不存在"

## 解决方案
- 将`db.First(&tag, tagID)`改为`db.First(&tag, uint(tagID))`
- 显式类型转换：`uint64` → `uint`
- 同时修复Get和Delete两个函数

## 影响范围
- GET /api/v1/tags/:id - 获取标签详情
- DELETE /api/v1/tags/:id - 删除标签

## 测试验证
- 标签ID 55（电子游戏）现在可以正常删除
- 删除后会自动刷新标签列表和计数

+++ backend/internal/handlers/tag.go
@@ -183,7 +183,7 @@ func (h *TagHandler) Get(c *gin.Context) {
 	db.Table("podcasts_tags").Where("tag_id = ?", tagID).Count(&count)
 
 	var tag models.Tag
-	if err := db.First(&tag, tagID).Error; err != nil {
+	if err := db.First(&tag, uint(tagID)).Error; err != nil {
 		c.JSON(http.StatusNotFound, gin.H{
 			"success": false,
 			"error": gin.H{
@@ -247,7 +247,7 @@ func (h *TagHandler) Update(c *gin.Context) {
 
 	// 检查标签是否存在
 	var tag models.Tag
-	if err := db.First(&tag, tagID).Error; err != nil {
+	if err := db.First(&tag, uint(tagID)).Error; err != nil {
 		c.JSON(http.StatusNotFound, gin.H{
 			"success": false,
 			"error": gin.H{
@@ -320,7 +320,7 @@ func (h *TagHandler) Delete(c *gin.Context) {
 
 	// 检查标签是否存在
 	var tag models.Tag
-	if err := db.First(&tag, tagID).Error; err != nil {
+	if err := db.First(&tag, uint(tagID)).Error; err != nil {
 		c.JSON(http.StatusNotFound, gin.H{
 			"success": false,
 			"error": gin.H{
